Add Server Support
Much of the code has been moved to main to make the mod usable on servers. The keybind for stepping forward one tick is a proxy for the command `/tick step`; it just sends the relevant command packet. As such, the same user permission level is required for the keybind as for calling the command through the chat.

The Command function was cleaned up slightly.